Split move validation/processing tests into a new test file
